一時払終身Ｐ計算（できた？）

+++ src/main/java/com/kohodx/insurancemathemticalservice/ichibarai_syushin/service/Ichibarai_SyusinCalcPremiumService.java
@@ -11,7 +11,7 @@ public class Ichibarai_SyusinCalcPremiumService {
     Ichibarai_SyusinCalcPremiumService ichibaraiSyusinCalcPremiumService;
 
     // 保険金（万単位）
-    private int insuredS;
+    private long insuredS;
     // 被保険者年齢
     private int insuredAge;
     private int Age;
